Small changes on weight name determination
The agreement coefficient functions in agree.coeff3.raw.r script file return the name of the weight set used (e.g. "unweighted", "quadratic", .etc...) When a numeric weight matrix is supplied it isn't always "custom weights." It could be identity weights and the weight name should reflect that.  This is the small problem I fixed.